Initial import
Initial implementation work for a system of types that also embeds/encodes
dimensional analysis information for use with physical quantities.

Beyond the core types, there is the beginning of a basic unit library in both
metric (`si` package) and US customary (`usc` package) systems.

There is also a simple expression builder that bridges the readability gap
caused by Go’s lack of operator overloading. It’s not perfect, but it’s a
start.

See `README.md` for more details and thoughts.